Add files via upload
A preliminary sweep of GLTC/WW Colligra and Clapur. This R script does many things: it backtracks to calculate the number of seeds from seed mass for GLTC Clapur, it also assumes that every open bloom for Colligra is perfect and has 4 seeds, instead of backtracking from empty inflorescence mass. It then cleans the data/removes NA's, adds 2 columns; one that is (Oseedstotal/Oblooms) - (Cseedstotal/Cblooms), and another column that is either "homogenous" or "heterogeneous" in regards to mix strategy. It then finally plots the seed difference column as a function of mix strategy. As of right, now there seems to be no sigifnicant difference in the magnitude of seed differences between the two mix strategy types. Graphs are at the bottom of the script.